+++ docs/versionhistory.rst
@@ -9,6 +9,8 @@ This library adheres to
 - Fixed ``TypeCheckError`` in unpacking assignment involving properties of a parameter
   of the function (`#506 <https://github.com/agronholm/typeguard/issues/506>`_;
   regression introduced in v4.4.1)
+- Fixed display of module name for forward references
+  (`#492 <https://github.com/agronholm/typeguard/pull/492>`_; PR by @JelleZijlstra)
 
 **4.4.1** (2024-11-03)
 
@@ -35,8 +37,6 @@ This library adheres to
   (`#465 <https://github.com/agronholm/typeguard/pull/465>`_)
 - Fixed basic support for intersection protocols
   (`#490 <https://github.com/agronholm/typeguard/pull/490>`_; PR by @antonagestam)
-- Fixed protocol checks running against the class of an instance and not the instance
-  itself (this produced wrong results for non-method member checks)
 
 **4.3.0** (2024-05-27)